refactor: mountain_id, name 가져오는 방법 수정
