--- application/usecases/admin/mountain/dto/AdminMountainNameDto.ts
@@ -1,4 +1,0 @@
-export interface AdminMountainNameDto {
-    mountain_id: number;
-    name: string;
-}
